Function to help deploy analysis folder inside a project folder

This function contains Rmd, R, md, files that help to structure an analysis following HCBC best-practices. For rnaseq, it will deploy: QC and DE Rmd with additional files to help to facilitate the analysis as needed.

Usage

bcbio_templates(type = "rnaseq", outpath = NULL, org = NULL)

Arguments

type

string indicating the type of analysis, supported: rnaseq.

outpath

string path indicating where to copy all the files to

org

string with the organization name. To deploy specific files.

Details

Normally these helper files are inside a report folder inside a project folder.

Examples

 if (FALSE) { # \dontrun{
  path <- withr::local_tempdir()
  bcbio_templates(type="base",outpath=path)
  fs::dir_ls(path,all=T)
 } # }